% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baseflow_Boughton}
\alias{baseflow_Boughton}
\title{compute a baseflow time series using the Boughton algorithm.}
\usage{
baseflow_Boughton(Q, k, C)
}
\arguments{
\item{Q}{numeric vector. Streamflow vector.}

\item{k}{float. parameter.}

\item{C}{float. parameter.}
}
\value{
Returns a baseflow time series.
}
\description{
This function compute baseflow time series using the algorithm of Boughton
described in Chapman (1999)
}
\references{
T. Chapman,
“A comparison of algorithms for streamflow recession and basefow separation,”
Hydrological Processes, vol. 13, pp. 701–714, 1999.
}
