% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baseflow_ChapmanMaxwell}
\alias{baseflow_ChapmanMaxwell}
\title{compute a baseflow time series using the Chapman & Maxwell algorithm.}
\usage{
baseflow_ChapmanMaxwell(Q, k)
}
\arguments{
\item{Q}{numeric vector. Streamflow vector.}

\item{k}{float. parameter.}
}
\value{
Returns a baseflow time series.
}
\description{
This function compute baseflow time series using the algorithm of Chapman and Maxwell
(1996)
}
\references{
T. G. Chapman and A. I. Maxwell,
“Baseflow separation - comparison of numerical methods with tracer experiments,” Hobart, Tasmania, 1996.
}
