% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baseflow_Eckhardt}
\alias{baseflow_Eckhardt}
\title{compute a baseflow time series using the Eckhardt algorithm.}
\usage{
baseflow_Eckhardt(Q, a, BFImax)
}
\arguments{
\item{Q}{numeric vector. Streamflow vector.}

\item{a}{float. parameter.}

\item{BFImax}{float. parameter.}
}
\value{
Returns a baseflow time series.
}
\description{
This function compute baseflow time series using the algorithm of Eckhardt (2005, 2008)
}
\references{
b K. Eckhardt,
“How to construct recursive digital filters for baseflow separation,”
Hydrological Processes, vol. 19, pp. 507–515, 2005.
K. Eckhardt,
“A comparison of baseflow indices, which were calculated with seven different baseflow separation methods,”
Journal of Hydrology, vol. 352, pp. 168–173, 2008, doi: 10.1016/j.jhydrol.2008.01.005.
}
