% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{baseflow_Gustard}
\alias{baseflow_Gustard}
\title{compute the baseflow time series using the Gustard algorithm.}
\usage{
baseflow_Gustard(Q, d = 5, k = 0.9)
}
\arguments{
\item{Q}{numeric vector. Streamflow vector.}

\item{d}{integer. Window size (in days) to look for local minima
(Default: 5).}

\item{k}{float. Factor to apply to local minima when looking for
pivot points (i.e. local local minima)}
}
\value{
Returns a baseflow time series.
}
\description{
This function compute baseflow time series using the algorithm of Gustard
et al. (XXXX) (a.k.a. method of the Institude of hydrology).
}
