% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{baseflow_LyneHollick}
\alias{baseflow_LyneHollick}
\title{compute a baseflow time series using the Lyne & Hollick algorithm.}
\usage{
baseflow_LyneHollick(Q, k)
}
\arguments{
\item{Q}{numeric vector. Streamflow vector.}

\item{k}{float. parameter.}
}
\value{
Returns a baseflow time series.
}
\description{
This function compute baseflow time series using the algorithm of Lyne and Hollick
(1979)
}
\references{
V. Lyne and M. Hollick,
“Stochastic time-variable rainfall-runoff modelling,”
in Hydrology and Water Ressources Symposium, Perth, Australia, 1979, pp. 89–92.
}
