% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{baseflow_magnitude}
\alias{baseflow_magnitude}
\title{Compute the baseflow regime magniture}
\usage{
baseflow_magnitude(Qbfr)
}
\arguments{
\item{Qbfr}{numeric vector. Baseflow regime vector resulting from the
\code{baseflow_regime()} function.}
}
\value{
A length 3-vector is return with the baseflow regime magnitude,
the min and the max of the baseflow regime
}
\description{
Compute the baseflow regime vector using the already computed baseflow
regime resulting from \code{baseflow_regime()} function
}
\seealso{
\link[=baseflow_regime]{baseflow_regime_magnitude}
}
