% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{baseflow_regime}
\alias{baseflow_regime}
\title{Compute the baseflow regime}
\usage{
baseflow_regime(Qbf, hdays, n = 1)
}
\arguments{
\item{Qbf}{numeric vector. Baseflow vector.}

\item{hdays}{numeric vector. Days of the (hydrological) year. see
get_hydro_years_seasons() function.}

\item{n}{integer. Window size for the rolling mean function}
}
\value{
A 366-long (or as many different hydrological days found in hdays
vector) numeric vector containing the baseflow regime values for each
calendar day.
}
\description{
Compute the baseflow regime vector i.e. the 'n'-days smoothed inter-annual
daily average of the baseflow time series
}
\seealso{
\link[=baseflow_magnitude]{baseflow_regime_magnitude}
}
