% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseflow.R
\name{baseflow_regime_magnitude}
\alias{baseflow_regime_magnitude}
\title{Compute the baseflow regime magnitude}
\usage{
baseflow_regime_magnitude(Qbf, hdays, n = 1)
}
\arguments{
\item{Qbf}{numeric vector. Baseflow vector.}

\item{hdays}{numeric vector. Days of the (hydrological) year. see
get_hydro_years_seasons() function.}

\item{n}{integer. Window size for the rolling mean function}
}
\value{
A length 3-vector is return with the baseflow regime magnitude,
the min and the max of the baseflow regime
}
\description{
Compute the baseflow magnitude i.e. the relative difference between the
maxima and minima of the 'n'-days smoothed inter-annual daily average of
the baseflow time series
}
\details{
A length 3-vector is return with the min and the max in second and third
position respectively
}
\seealso{
\link[=baseflow_magnitude]{baseflow_regime}
}
