% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{bias}
\alias{bias}
\title{compute the bias between simulated and observed data}
\usage{
bias(obs, sim, na.rm = TRUE, sim_minus_obs = TRUE)
}
\arguments{
\item{obs}{numeric vector. Observed streamflow vector.}

\item{sim}{numeric vector. Simulated streamflow vector.}

\item{na.rm}{logical. Whether missing values should be ignored.}

\item{sim_minus_obs}{logical.  should it be sim - obs? (or the other
way around)}
}
\value{
Returns a float value corresponding to the bias.
}
\description{
This function compute the bias (unitless) between simulated and observed
streamflow data.
}
