% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cumsum_interpolate}
\alias{cumsum_interpolate}
\title{Wrapper around the \code{cumsum} function where missing values can be
interpolated.}
\usage{
cumsum_interpolate(x, na.action = "interpolate")
}
\arguments{
\item{x}{vector to use to compute cumulated values}

\item{na.action}{action to undertake when missing values are found (default:
'interpolate', any other value leads to the original \code{cumsum} function)}
}
\value{
the cumulative sum of x.
}
\description{
Wrapper around the \code{cumsum} function where missing values can be
interpolated.
}
