% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cumulative_curves}
\alias{cumulative_curves}
\title{Compute cumulative curves on a data.frame of matrix}
\usage{
cumulative_curves(x, na.action = "interpolate")
}
\arguments{
\item{x}{numeric vector, data.frame or matrix. a vector (of length m) or
a matrix/data.frame (with m rows)}

\item{na.action}{character. a character string indicating whether missing
values should be interpolated or not (see \code{cumsum_interpolate()})}
}
\value{
A data.frame with 'm' rows containing the cumulative version of
'x'.
}
\description{
Given a vector or a matrix/data.frame 'x', compute the corresponding
cumulative curve. This function is only a wrapper around
\code{cumsum_interpolate()} with a call to \code{apply()}.
}
