% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_duration_curve.R
\name{fdc_percentiles}
\alias{fdc_percentiles}
\title{Compute any flow duration curve percentiles}
\usage{
fdc_percentiles(Q, probs = c(0.1, 0.9))
}
\arguments{
\item{Q}{numeric vector. Streamflow values.}

\item{probs}{numeric vector containing the exceedance probabilities for the
streamflow percentiles to compute.}
}
\value{
A numeric vector of the same length as \code{probs} containing the
streamflow percentiles corresponding to the exceedance probabilities
specified with argument \code{probs}.
}
\description{
Given streamflow exceedance probability values (\code{probs}) compute the
corresponding streamflow values/percentiles.
}
\seealso{
\code{\link{fdc_slope}} and \code{\link{fdc_slope_percentiles}}
}
