% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_duration_curve.R
\name{fdc_slope}
\alias{fdc_slope}
\title{Compute the slope of the mid-segment slope of the flow duration curve}
\usage{
fdc_slope(Q, probs = c(0.33, 0.66))
}
\arguments{
\item{Q}{numeric vector. Streamflow values.}

\item{probs}{numeric vector of length 2 containing the exceedance
probabilities that define the start and end of the mid-segment of the
flow duration curve.}
}
\value{
A single numeric value corresponding to the slope of the mid-segment
of the flow duration curve.
}
\description{
Given two probability of streamflow exceedance (probs) which define the
start and end of the mid-segment of the flow duration curve, compute the
slope of the mid-segment slope of the flow duration curve.
}
\details{
The computation of the slope is done with the following function:
\code{- (log10(Q_high) - log10(Q_low)) / diff(probs)}
where \code{log10(Q_low)} (resp. \code{log10(Q_high)}) is the log transformed
values of streamflow corresponding to the low (resp. high) end of the
mid-segment of the flow duration curve
}
\seealso{
\code{\link{fdc_percentiles}} and \code{\link{fdc_slope_percentiles}}
}
