% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_duration_curve.R
\name{fdc_slope_percentiles}
\alias{fdc_slope_percentiles}
\title{Compute flow duration curve based hydrological signatures}
\usage{
fdc_slope_percentiles(
  Q,
  slope_probs = c(0.33, 0.66),
  percentile_probs = c(0.1, 0.9)
)
}
\arguments{
\item{Q}{numeric vector. Streamflow values.}

\item{slope_probs}{numeric vector of length 2 containing the exceedance
probabilities that define the start and end of the mid-segment of the
flow duration curve.}

\item{percentile_probs}{numeric vector containing the exceedance
probabilities for the streamflow percentiles to compute}
}
\value{
A numeric vector of length n+1 where n is the length of
'percentile_probs' with the first element being the slope of the
flow duration curve mid-segment.
}
\description{
This function compute both (1) the mid-segment flow duration curve slope
(see \code{\link{fdc_slope}}) and (2) any streamflow quantile (see
\code{\link{fdc_percentiles}}).
}
\seealso{
\code{\link{fdc_slope}} and \code{\link{fdc_percentiles}}
}
