% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_valid_hydro_year}
\alias{get_valid_hydro_year}
\title{Compute the valid hydrological year}
\usage{
get_valid_hydro_year(hy, x = NULL, n = 365, na.th = 0.05)
}
\arguments{
\item{hy}{numeric vector. a hydrological year vector (will be coerced to factor)}

\item{x}{numeric matrix. a matrix (or a vector) with as many row (elements)
as the length of \code{hy} used to look for missing values}

\item{n}{integer. the minimum length of a complete hydrological year this is
also use to compute the proportion of missing values.}

\item{na.th}{numeric. the minimal tolerated proportion of missing values}
}
\value{
a logicalvector of the same length as the hydrological year vector
that indicates whether or not the time steps are part of a valid hydrological
year.
}
\description{
Given a hydrological year vector and a matrix with the same number
of rows (and any number of column), this function returns a logical
vector of the same length as the hydrological year vector that
indicates whether or not the time steps are part of a valid
hydrological year. Validity of a hydrological year is assessed
according the two following rules:
\enumerate{
\item is the year complete (i.e. is there at least \code{n} (365) days)?
\item is there less than \code{na.th} (proportion) missing values in the year?
}
}
