% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_what_is_within}
\alias{get_what_is_within}
\title{What is within?}
\usage{
get_what_is_within(OriginalTime, FilteredTime)
}
\arguments{
\item{OriginalTime}{vector. first time/date vector}

\item{FilteredTime}{vector. second time/date vector}
}
\value{
logical vector of the size of the
first vector indicating whether or not the times/dates were found in the
second vector
}
\description{
Given two vectors (typically date vectors), return a logical vector of the
size of the first vector indicating whether or not the times/dates were
found in the second vector
}
