% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{HSA}
\alias{HSA}
\title{HSA (Hydrological Signature Analysis)}
\description{
This package is a set of function to compute hydrological signatures
from streamflow, precipitation and air temperature time series.
}
\author{
Ivan Horner
}
