% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{kge}
\alias{kge}
\title{Computes the Kling-Gupta efficiency coefficient.}
\usage{
kge(obs, sim, na.rm = TRUE, method = 1)
}
\arguments{
\item{obs}{numeric vector. Observed streamflow vector.}

\item{sim}{numeric vector. Simulated streamflow vector.}

\item{na.rm}{logical. Whether missing values should be ignored.}

\item{method}{two methods are implemented (see hydroGOF::KGE): one from
Gupta et al. (2009) and the other one from Kling et al. (2012)}
}
\value{
Returns a float value between -Inf and 1 corresponding to the
Kling-Gupta efficiency coefficient.
}
\description{
This function computes the Kling-Gupta efficiency coefficient. It islargely
inspired by the similar function in the HydroGOF package. Here, only a
simplified version is provided with no checks on inputs and very little
formatting.
}
