% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{log_transform}
\alias{log_transform}
\title{Computes log transformed streamflow vector.}
\usage{
log_transform(x, method = "Pushpalatha2012")
}
\arguments{
\item{x}{numeric vector. Streamflow vector to log transform.}

\item{method}{Method used to deal with missing values: 'Pushpalatha2012'
a numeric value (to be added to 'x') or 'inf.na'. If a charcter string
equal to "Pushpalatha2012", the method of Pushpalatha et al. (2012) is used.
If a character string equal to "inf.na", any non-finite value are set to NA.
If a flow value, it is added to each streamflow value before transofmation.}
}
\value{
Returns a numeric vector of the same length than x containing the log
transformed streamflow values.
}
\description{
compute the log transformed value of streamflow using either the approach
proposed by Pushpalatha et al. (2012) (i.e. a 100th of the mean value is
added to each streamflow value), or by adding any given value to all
streamflow values or simply by setting to NA any non finite value resulting
from the application of the log() function.
}
