% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mct_approach.R
\name{mct_inflexionpoints}
\alias{mct_inflexionpoints}
\title{Compute the inflexion point of the streamflow mass curve}
\usage{
mct_inflexionpoints(
  Qcum,
  periods = c(100, 300),
  n = 30L,
  return_derivative = FALSE
)
}
\arguments{
\item{Qcum}{numeric vector. Cumulative streamflow regime vector.}
}
\value{
If \code{return_derivative} is TRUE, a list with 3 components is
returned:
\itemize{
\item \code{data}: a data.frame containing two column: the derivative of \code{Qr}
(dQ) and its smoothed version (dQs).
\code{inflexion_points} the indices of the two inflexions points found
If \code{return_derivative} is FALSE, only the two inflexions points are
returned in a length 2 numeric vector.
}
}
\description{
From a streamflow regime vector \code{Qr} (i.e. a inter-annual average of
streamflow for each calendar day), it searches for the inflexion points in
a given time window \code{p}. This functions (1) computes the derivative of
\code{Qr}, (2) and smooth it using a rolling mean with a window \code{n}.
}
\seealso{
\link{mct_snowstorage}
}
