% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mct_approach.R
\name{mct_snowstorage}
\alias{mct_snowstorage}
\title{Compute snow storage estimate using the MCT technique}
\usage{
mct_snowstorage(
  Pcum,
  Qcum,
  inflexion_points,
  use_tangentes = FALSE,
  period = c(1, length(Pcum)),
  fixed_intercept = FALSE,
  return_tangents = FALSE
)
}
\arguments{
\item{Pcum}{numeric vector. Cumulative precipitation regime curve.}

\item{Qcum}{numeric vector. Cumulative streamflow regime curve.}

\item{inflexion_points}{numeric vector. A length 2 numeric vector containing
the two inflexion points of the cumulative streamflow regime curve (see
\code{mct_inflexionpoints()})}

\item{use_tangentes}{logical. Should the methodology of Schaefli (2016) be used
to compute two snow estimates (see details).}

\item{fixed_intercept}{logical. Should a fixed intercept at (0, 0) be used
in the linear regression (see details).}

\item{return_tangents}{logical. Should the tangents values be returned as well.}

\item{period.}{numeric vector. A length 2 numeric vector defining the period
to use to perform the linear regression on \code{Pcum} (see details).}
}
\value{
There are three cases:
\enumerate{
\item \code{use_tangentes} is FALSE: a single numeric value corresponding to one
storage estiamte.
\item \code{use_tangentes} is TRUE and \code{return_tangents} is FALSE: a length 2
numeric vector is returned containing 2 snow storage estimates: (1) from the
accumulation period (Sa) and (2) from the  metling period (Sm).
\item \code{use_tangentes} is TRUE and \code{return_tangents} is TRUE: a list is
returned containing the the two snow storage estimates (S), the slope of resulting
from the linear regression over \code{Pcum} (P_slp), the intercept and slope of the
the two tangents to streamflow at inflexion point 1 (Qa_slp) and inflexion point 2
(Qm_slp). These last two components are length 2 numeric vectors.
}
}
\description{
Given cumulative regime curve of precipitation \code{Pcum} and streamflow
\code{Qcum}, the inflexions points of the streamflow regime curve returned
by the function \code{mct_inflexionpoints()}, this function compute the
one or two snow storage estimates depending on the method used (see details).
}
\details{
If \code{use_tangentes} is FALSE, the snow storage estimate is simply the
difference \code{Pcum[inflexion_points[1L]] - Qcum[inflexion_points[1L]]}.
If \code{use_tangentes} is TRUE, a linear regression is performed on
Pcum to retrieve the average slope in the period defined by \code{period}.
If \code{fixed_intercept} is TRUE, a fixed intercept at (0, 0) is used
in the linear regression. If \code{use_tangentes} is TRUE, two snow estimates
are computed: (1) from the accumulation period (Sa) and (2) from the
metling period (Sm). For details see Schaefli (2016)
}
\references{
B. Schaefli, “Snow hydrology signatures for model identification within a limits-of-acceptability approach”,
Hydrological Processes, vol. 30, no. 22, Art. no. 22, Aug. 2016, doi: 10.1002/hyp.10972.
}
\seealso{
\link{mct_snowstorage}
}
