% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{nse}
\alias{nse}
\title{Computes the Nash-Sutcfliff efficiency coefficient}
\usage{
nse(obs, sim, na.rm = TRUE)
}
\arguments{
\item{obs}{numeric vector. Observed streamflow vector.}

\item{sim}{numeric vector. Simulated streamflow vector.}

\item{na.rm}{logical. Whether missing values should be ignored.}
}
\value{
Returns a float value between -Inf and 1 corresponding to the
Nash-Sutcfliff efficiency coefficient.
}
\description{
Given two streamflow vectors of same length (observer 'obs' and simulated
'sim' streamflow), this functions computes the corresponding Nash-Sutcliff
efficiency.
}
