% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{nselog}
\alias{nselog}
\title{Computes the Nash-Sutcfliff efficiency coefficient on the log transformed
streamflow values}
\usage{
nselog(obs, sim, na.rm = TRUE, log_method = "inf.na")
}
\arguments{
\item{obs}{numeric vector. Observed streamflow vector.}

\item{sim}{numeric vector. Simulated streamflow vector.}

\item{na.rm}{logical. Whether missing values should be ignored.}

\item{log_method}{character string or numeric. See log_transform().}
}
\value{
Returns a float value between -Inf and 1 corresponding to the
Nash-Sutcfliff efficiency coefficient computed on the log transofmed
streamflow values.
}
\description{
This function computes the Nash-Sutcliff efficiency coefficient using the
nse() function on the streamflow values which are first log transformed
using the log_transform()vfunction.
}
