% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_approach.R
\name{pq_cumsum}
\alias{pq_cumsum}
\title{Computes the cumulative inter-annual daily average of P and Q as
well as the difference P-Q.}
\usage{
pq_cumsum(Q, P, hdays)
}
\arguments{
\item{Q}{numeric vector. Streamflow values.}

\item{P}{numeric vector. Precipitation values.}

\item{hdays}{numeric vector. Days of the (Hydrological) years vector}
}
\value{
A 3-columns data.frame is returned (Q, P and P_Q) containing the
cumulative curve of streamflow (Q), precipitation (P) and the differences
(P-Q)
}
\description{
Computes the cumulative inter-annual daily average of P and Q as
well as the difference P-Q.
}
