% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_approach.R
\name{pq_slopes}
\alias{pq_slopes}
\title{P-Q signatures computations using Q, P and hdays vectors}
\usage{
pq_slopes(
  Q,
  P,
  hdays,
  start = 15,
  end = 183,
  bp = mean(c(start, end)),
  intercept = TRUE
)
}
\arguments{
\item{Q}{numeric vector. Streamflow values.}

\item{P}{numeric vector. Precipitation values.}

\item{hdays}{numeric vector. Days of the (Hydrological) years vector}

\item{start}{numeric. The day number of the start of the period to
search for a change of trend in P-Q.}

\item{end}{numeric. The day number of the end of the period to
search for a change of trend in P-Q.}

\item{bp}{numeric value. The day number of the initial guess of threshold
date}

\item{intercept}{logical. Should the intercept be estimated (default: TRUE)
or fixed to c(0, 0) (FALSE)?}
}
\value{
Returned
}
\description{
It computes the cumulative inter-annual daily average of P and Q as
well as the difference P-Q. Then, it computes the seasonal response change
signatures following the so-called P-Q approach: breackpoint date,
first period slope (dry), second period slope (wet) and the intercepts
associated with these two slopes (only there for plotting purposes).
}
\details{
This function uses a segmented regression to find the dry and wet slopes.
It also returns the intercepts of the two lines (which are not signatures)
that can be used to plot the lines.
}
