% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pq_approach.R
\name{pq_snowstorage}
\alias{pq_snowstorage}
\title{Compute P-Q snow storage estimate}
\usage{
pq_snowstorage(Pcum, Qcum, period = NULL, return_data = FALSE)
}
\arguments{
\item{Pcum}{Numeric vector. Cumulative precipitation over one hydrological
year (or inter-annual average hydrological year).}

\item{Qcum}{Numeric vector. Cumulative streamflow over one hydrological
year (or inter-annual average hydrological year).}

\item{period}{NULL or a numeric vector of length 2 that defines the end and
start (number of days) of the period where the maximum of the P-Q curve
is to be found.}

\item{return_data}{logical. Should the intermediate results (P-Q curve and
number of the day corresponding to the maximum of the P-Q curve) be also
returned? Default is FALSE. and a single value corresponding to the snow
storage estimate is returned. If TRUE, a list with the snow storage estimate
('S'), and the P-Q curve ('PQ') and the date of the maximum of the P-Q curve
('tQcumMax') is returned.}
}
\value{
If return_data is FALSE, a single value corresponding to the snow
storage estimate is returned. Otherwise, a list with the snow storage
estimate ('S'), and the P-Q curve ('PQ') and the date of the maximum of
the P-Q curve ('tQcumMax') is returned.
}
\description{
Given the cumulative curve of precipitation (Pcum) and streamflow (Qcum), it
computes the differences (Pcum - Qcum) and computes a snow storage estimate
corresponding to the maximum of the Pcum - Qcum curve.
}
