% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_regime_signatures.R
\name{q_timing}
\alias{q_timing}
\title{Compute streamflow regime maximum timing}
\usage{
q_timing(Q)
}
\arguments{
\item{Q}{numeric vector. streamflow regime (365 or 366 values)}
}
\value{
the index of the maximum of input 'Q'
}
\description{
Compute the timing of the maximum of the streamflow regime. This function
is only a wrapper arount which.max() doing ... nothing.
}
