% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recessions_analysis.R
\name{rec_power_law_model_fit}
\alias{rec_power_law_model_fit}
\title{Computes the recession time and recession concavity of a recession event.}
\usage{
rec_power_law_model_fit(x)
}
\arguments{
\item{x}{numeric vector. Vector containing the streamflow value for
a single recesion event.}
}
\description{
This function fits a power law model ((1/tau)*Q^b) to a single recession provided
as numeric vector. It returns the computed recession time 'tau' and concavity 'b'.
}
