% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recessions_analysis.R
\name{recession_events_index}
\alias{recession_events_index}
\title{Extract recession events from a streamflow time series}
\usage{
recession_events_index(
  Q,
  minduration,
  maxduration,
  minvalue,
  n_smooth,
  window_smooth,
  return_smoothed = FALSE
)
}
\arguments{
\item{Q}{numeric vector. Streamflow vector.}

\item{minduration}{integer. Minimum duration of a recession event in days}

\item{maxduration}{integer. Maximum duration of a recession event in days}

\item{minvalue}{float. Minimum streamflow value to consider a local maximum
as a potential start of a recession event}

\item{n_smooth}{integer. How many times the rolling mean is to be applied
on the streamflow timeseries before applying the recession events extraction}

\item{window_smooth}{integer. The size, in days, of the window over which
the rollogin mean is to be applied.}

\item{return_smoothed}{logical. Whether the smoothed streamflow timeseries
should be returned as well.}
}
\value{
Returns a two columns data.frame with the 'start' and 'end' index
of each recession event. If 'return_smoothed' is TRUE, a list with two
elements is returned: 'i_events' containing the the data.frame with the
'start' and 'end' index of the recession events and 'Q_smoothed' containing
the smoothed streamflow vector.
}
\description{
This function extracts the index of the start and end of each recession
event which are returned in a two columns data.frame
}
