% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{regime_curve}
\alias{regime_curve}
\title{Compute a regime curve}
\usage{
regime_curve(hdays, x, n = 1L, na.rm = TRUE, remove_last = TRUE)
}
\arguments{
\item{hdays}{numeric vector. vector of 'julian' date of length m which
should be computed according to hydrological year (see function
\code{get_hydro_years_seasons()})}

\item{x}{numeric vector. a vector (of length m) or a matrix (with m rows)}

\item{n}{integer. smoothing window}

\item{remove_last}{logical. should the last day of the year (366) be
removed (recommended).}
}
\value{
A 366 (remove_last = FALSE) or 365 (remove_last = TRUE) long
numeric vector (or matrix with 366 or 365 rows) containing the regime curve(s)
}
\description{
Compute regime curves from the data provided in 'x'. It first apply
smoothing (rolling mean) with a wondow 'n' of the data provided in 'x'.
Then, it compute the inter-annual mean of each calendar day (provided in
vector 'hdays'). It recommended to ignore the last day using
'remove_last = TRUE'.
}
