% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_metrics.R
\name{rsquare}
\alias{rsquare}
\title{compute the coefficient of determination between simulated and observed data}
\usage{
rsquare(obs, sim, na.rm = TRUE, method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{obs}{numeric vector. Observed streamflow vector.}

\item{sim}{numeric vector. Simulated streamflow vector.}

\item{na.rm}{logical. Whether missing values should be ignored.}

\item{method}{character. see cor() function.}
}
\value{
Returns a float value corresponding to the coefficient of
determination.
}
\description{
This function compute the compute the coefficient of determination using
the cor() function between simulated and observed data
}
