% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runoff_coefficient.R
\name{runoff_coefficient}
\alias{runoff_coefficient}
\title{Compute the runoff coefficient}
\usage{
runoff_coefficient(Q, P, na.rm = TRUE)
}
\arguments{
\item{Q}{numeric vector. Streamflow values.}

\item{P}{numeric vector. Precipitation values.}

\item{na.rm}{logical. Should missing values be omited?}
}
\value{
A single numeric value corresponding to the runoff coefficient
}
\description{
compute the runoff coefficient or ratio: the ratio between the total
strealfow volume and the total precipitaion volume.
}
\details{
Warning: no checks are done on inputs! You should have selected the proper
period beforehand: both vector should be of same length and span over the
same period. In addition the runoff coefficient should be compute from the
start of a hydrological year to the end of a hydrological year.
}
