% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sum_ratio}
\alias{sum_ratio}
\title{Compute the ratio of two sums.}
\usage{
sum_ratio(top, bottom, na.rm = TRUE)
}
\arguments{
\item{top}{numeric vector. Its sum will be the nominator of the ratio}

\item{bottom}{numeric vector. Its sum will be the denominator of the ratio}

\item{na.rm}{logical. Should missing values be omited?}
}
\value{
A single numeric value corresponding to the ratio of the sum of
'top' and the sum of 'bottom'
}
\description{
Compute the ratio between the sum of 'top' and the sum of 'bottom'.
}
