% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_regime_signatures.R
\name{tq_periods}
\alias{tq_periods}
\title{Compute temperature-streamflow regime cycle periods}
\usage{
tq_periods(T, Q, periods = c(183, 336, 62))
}
\arguments{
\item{T}{numeric vector. air temperature regime (365 or 366 values)}

\item{Q}{numeric vector. streamflow regime (365 or 366 values)}

\item{peridos}{numeric vector. a vector of length 3 defining (1) the start
of the T+Q+ period, (2) the start (resp. end) of the T-Q-2 (resp. T-Q-1)
period and (3) the end of the T-Q-2 period. Note: the other period
boundaries are computed using the air temperature and streamflow regime
maxima timings.}
}
\value{
A list containing the indices of the different periods in 4
components: "TpQp", "TpQm", "TmQm1", "TmQm2"
}
\description{
Compute list containing the indices of each periods necessary to compute
the slopes from the T-Q regime cycle:
\itemize{
\item T+Q+ (TpQp) period
\item T+Q- (TpQm) period
\item T-Q-1 (TmQm1) period
\item T-Q-2 (TmQm2) period
}
}
\references{
I. Horner, F. Branger, H. McMillan, O. Vannier, and I. Braud,
“Information content of snow hydrological signatures based on streamflow,
precipitation and air temperature,”
Hydrological Processes, vol. 34, no. 12, Art. no. 12, 2020, doi: 10.1002/hyp.13762.
}
