% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_regime_signatures.R
\name{tq_signatures}
\alias{tq_signatures}
\title{Compute signatures based on the air temperature (T) and streamflow (Q) regime}
\usage{
tq_signatures(
  hdays,
  T,
  Q,
  n = 30,
  remove_last = TRUE,
  periods = c(183, 336, 62)
)
}
\arguments{
\item{hdays}{numeric vector. vector of 'julian' date of length m which
should be computed according to hydrological year (see function
\code{get_hydro_years_seasons()})}

\item{T}{numeric vector. air temperature regime (365 or 366 values)}

\item{Q}{numeric vector. streamflow regime (365 or 366 values)}

\item{n}{integer. smoothing window}

\item{remove_last}{logical. should the last day of the year (366) be}

\item{peridos}{numeric vector. a vector of length 3 defining (1) the start
of the T+Q+ period, (2) the start (resp. end) of the T-Q-2 (resp. T-Q-1)
period and (3) the end of the T-Q-2 period. Note: the other period
boundaries are computed using the air temperature and streamflow regime
maxima timings.}
}
\value{
A vector of length 5 with the following named components:
"TQslp_TpQp", "TQslp_TpQm", "TQslp_TmQm1", "TQslp_TmQm2", "tQmax"
}
\description{
This function computes signatures based on the air temperature (T) and
streamflow (Q) regime cycle: the slopes of 4 periods (T+Q+, T+Q-, T-Q-1,
T-Q-2  periods) of the T-Q cycle and timing of the Q maximum.
}
\references{
I. Horner, F. Branger, H. McMillan, O. Vannier, and I. Braud,
“Information content of snow hydrological signatures based on streamflow,
precipitation and air temperature,”
Hydrological Processes, vol. 34, no. 12, Art. no. 12, 2020, doi: 10.1002/hyp.13762.
}
\seealso{
[tq_periods][regime_curve][q_timing][tq_slopes]
}
