% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tq_regime_signatures.R
\name{tq_slopes}
\alias{tq_slopes}
\title{Compute the temperature-streamflow regime cycle slopes}
\usage{
tq_slopes(T, Q, indices)
}
\arguments{
\item{T}{numeric vector. air temperature regime (365 or 366 values)}

\item{Q}{numeric vector. streamflow regime (365 or 366 values)}

\item{indices}{list. list of  length 'n'  containing the vectors of indices
defining the different period to use to compute the slopes/intercept in
\code{T = b + slp * Q}}
}
\value{
A matrix with 'n' rows (the different periods) and 2 columns: 'intercept'
and 'slope'
}
\description{
Compute slopes (and intercepts) of each periods from the air
temperature - streamflow regime cycle using the periods defined in a list of
indices vectors (see \code{tq_periods()})
}
\references{
I. Horner, F. Branger, H. McMillan, O. Vannier, and I. Braud,
“Information content of snow hydrological signatures based on streamflow,
precipitation and air temperature,”
Hydrological Processes, vol. 34, no. 12, Art. no. 12, 2020, doi: 10.1002/hyp.13762.
}
\seealso{
\link{tq_periods}
}
